[exporter/signalfx_correlation] Move correlation from sapm to signalfx_correlation
