+++ exporter/sapmexporter/factory.go
@@ -16,13 +16,10 @@ package sapmexporter
 
 import (
 	"context"
-	"time"
 
-	"github.com/signalfx/signalfx-agent/pkg/apm/correlations"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
-	"go.opentelemetry.io/collector/translator/conventions"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/common/splunk"
 )
@@ -53,22 +50,6 @@ func createDefaultConfig() configmodels.Exporter {
 		TimeoutSettings: exporterhelper.CreateDefaultTimeoutSettings(),
 		RetrySettings:   exporterhelper.CreateDefaultRetrySettings(),
 		QueueSettings:   exporterhelper.CreateDefaultQueueSettings(),
-		Correlation: CorrelationConfig{
-			Enabled:             false,
-			StaleServiceTimeout: 5 * time.Minute,
-			SyncAttributes: map[string]string{
-				conventions.AttributeK8sPodUID:   conventions.AttributeK8sPodUID,
-				conventions.AttributeContainerID: conventions.AttributeContainerID,
-			},
-			Config: correlations.Config{
-				MaxRequests:     20,
-				MaxBuffered:     10_000,
-				MaxRetries:      2,
-				LogUpdates:      false,
-				RetryDelay:      30 * time.Second,
-				CleanupInterval: 1 * time.Minute,
-			},
-		},
 	}
 }
