+++ exporter/sapmexporter/factory_test.go
@@ -38,7 +38,6 @@ func TestCreateExporter(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	eCfg := cfg.(*Config)
 	eCfg.Endpoint = "http://local"
-	eCfg.Correlation.Endpoint = "http://local"
 	params := component.ExporterCreateParams{Logger: zap.NewNop()}
 
 	te, err := factory.CreateTraceExporter(context.Background(), params, eCfg)
@@ -48,15 +47,4 @@ func TestCreateExporter(t *testing.T) {
 	me, err := factory.CreateMetricsExporter(context.Background(), params, eCfg)
 	assert.Error(t, err)
 	assert.Nil(t, me)
-}
-
-func TestCreateExporterWithoutAPIEndpoint(t *testing.T) {
-	factory := NewFactory()
-	cfg := factory.CreateDefaultConfig().(*Config)
-	cfg.Endpoint = "http://local"
-	cfg.Correlation.Enabled = true
-	params := component.ExporterCreateParams{Logger: zap.NewNop()}
-	te, err := factory.CreateTraceExporter(context.Background(), params, cfg)
-	assert.Nil(t, te)
-	assert.EqualError(t, err, "`correlation.endpoint` must be set when `correlation.enabled` is true")
 }
