+++ exporter/signalfxcorrelationexporter/spanshims.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sapmexporter
+package signalfxcorrelationexporter
 
 import (
 	"github.com/signalfx/signalfx-agent/pkg/apm/tracetracker"
